Models updates. Cal event generation
