+++ backend/news/models.py
@@ -1,8 +1,7 @@
 from core.constants import FieldLength
 from core.models import TimeStamp
-from data.models import Disciplines
 from django.db import models
-from info.models import RegionalDivisions
+from info.models import Disciplines, RegionalDivisions
 from users.models import Users
 
 
@@ -11,7 +10,9 @@ class News(TimeStamp):
         max_length=FieldLength.MAX_LENGTH_TITLE,
         verbose_name="Название новости",
     )
-    news_image = models.ImageField(upload_to="images/news/", verbose_name="Фотография")
+    news_image = models.ImageField(
+        upload_to="images/news/", verbose_name="Фотография к новости"
+    )
     news_text = models.TextField(verbose_name="Описание новости")
     news_regional_divisions = models.ManyToManyField(
         RegionalDivisions, verbose_name="Региональные отделения", related_name="news"
